Move HomeViewModelPreview
