+++ .idea/misc.xml
@@ -24,6 +24,8 @@
         <entry key="../../../../layout/compose-model-1639923666491.xml" value="0.14629629629629629" />
         <entry key="../../../../layout/compose-model-1639925045918.xml" value="0.14629629629629629" />
         <entry key="../../../../layout/compose-model-1639925153829.xml" value="0.67" />
+        <entry key="../../../../layout/compose-model-1639927318896.xml" value="0.17246437552388935" />
+        <entry key="../../../../layout/compose-model-1639927649285.xml" value="0.17246437552388935" />
       </map>
     </option>
   </component>
