+++ feature/home/src/main/java/com/tfandkusu/template/view/home/listitem/GitHubRepoListItem.kt
@@ -153,7 +153,6 @@ fun LanguageLabel(language: String) {
     }
 }
 
-
 @Composable
 @Preview
 fun GitHubRepoListItemPreviewNormal() {
